+++ src/rest-apis/blockchain/ETH/token.js
@@ -45,6 +45,37 @@ class EthToken extends EthBaseToken {
         return this.request.get(this.basePath + this.getSelectedNetwork() + '/tokens/address/' + address, combinedQueryParams);
     }
 
+    /**
+     * Estimate Token Transaction GasLimit
+     *
+     * @async
+     * @desc Crypto APIs provides an endpoint for estimating the gas limit needed for a successful transaction.
+     *
+     * @param {string} fromAddress - From address.
+     * @param {string} toAddress - To address.
+     * @param {string} contract - Contract address.
+     * @param {number} tokenAmount - Token amount of the transaction.
+     * @param {object} [optData] - Optional data.
+     * @param {object} [queryParams] - Additional query parameters.
+     */
+    estimateTokenTransactionGasLimit(fromAddress, toAddress, contract, tokenAmount, optData = {}, queryParams = {}) {
+        let data = {};
+
+        Object.keys(optData).map(k => {
+            data[k] = optData[k];
+        });
+
+        data = {
+            ...data,
+            fromAddress: fromAddress,
+            toAddress: toAddress,
+            contract: contract,
+            tokenAmount: tokenAmount
+        };
+
+        return this.request.post(this.basePath + this.getSelectedNetwork() + '/tokens/transfer/gas-limit', data, queryParams);
+    }
+
 }
 
 module.exports = EthToken;