+++ src/rest-apis/blockchain/XRP/index.js
@@ -1,32 +1,24 @@
 const BlockchainNetwork = require('../../../common/blockchain/blockchain-network');
 
 const Blockchain = require('../../../common/blockchain/base-blockchain');
+const Webhook = require('../../../common/blockchain/base-web-hook');
 
-const Address = require('../../../common/blockchain/eth-base-address');
-const PaymentForwarding = require('../../../common/blockchain/eth-base-payment-forwarding');
-
-const Token = require('./token');
-const Contract = require('./contract');
+const Address = require('./address');
 const Transaction = require('./transaction');
-const Webhook = require('./webhook');
 
-const ID = 'eth';
+const ID = 'xrp';
 const NETWORKS = {
     MAINNET: 'mainnet',
-    ROPSTEN: 'ropsten',
-    RINKEBY: 'rinkeby',
+    TESTNET: 'testnet',
 };
 
-class ETH extends BlockchainNetwork {
+class XRP extends BlockchainNetwork {
 
     constructor(...props) {
         super(...props, ID, NETWORKS, NETWORKS.MAINNET);
 
         this.address = new Address(...props, ID);
         this.blockchain = new Blockchain(...props, ID);
-        this.contract = new Contract(...props, ID);
-        this.paymentForwarding = new PaymentForwarding(...props, ID);
-        this.token = new Token(...props, ID);
         this.transaction = new Transaction(...props, ID);
         this.webhook = new Webhook(...props, ID);
     }
@@ -34,6 +26,6 @@ class ETH extends BlockchainNetwork {
 }
 
 module.exports = {
-    ETH,
+    XRP,
     NETWORKS,
 };