+++ tests/rest-apis/blockchains/BTC/webhook.test.js
@@ -8,14 +8,26 @@ async function Webhook(caClient) {
     const omniTestTx = 'e11b979a9396282b3aa53f57512fa187a820069ae07259c969589246151198d6';
     const omniTestAddress = 'mittVBKCvyVRNVc5cq5sWXf8VsLsvn9j4f';
 
-    await caClient.BC.BTC.webhook.createNewBlockWebHook(testUrl).then(response => response.payload);
-    await caClient.BC.BTC.webhook.createConfirmedTransactionWebHook(testUrl, testTx, 1).then(response => response.payload);
-    await caClient.BC.BTC.webhook.createAddressTransactionWebHook(testUrl, testAddress, 1).then(response => response.payload);
-    await caClient.BC.BTC.webhook.createTransactionConfirmationsWebHook(testUrl, testAddress, 1).then(response => response.payload);
+    await caClient.BC.BTC.webhook.createNewBlockWebHook(testUrl);
+    await caClient.BC.BTC.webhook.createNewBlockWebHook(testUrl, {allowDuplicate: true});
 
-    await caClient.BC.BTC.webhook.createOmniConfirmedTransactionWebHook(testUrl, omniTestTx, 1).then(response => response.payload);
-    await caClient.BC.BTC.webhook.createOmniTransactionConfirmationsWebHook(testUrl, omniTestAddress, 1).then(response => response.payload);
-    await caClient.BC.BTC.webhook.createOmniTokenWebHook(testUrl, omniTestAddress, 1).then(response => response.payload);
+    await caClient.BC.BTC.webhook.createConfirmedTransactionWebHook(testUrl, testTx, 1);
+    await caClient.BC.BTC.webhook.createConfirmedTransactionWebHook(testUrl, testTx, 1, {allowDuplicate: true});
+
+    await caClient.BC.BTC.webhook.createAddressTransactionWebHook(testUrl, testAddress, 1);
+    await caClient.BC.BTC.webhook.createAddressTransactionWebHook(testUrl, testAddress, 1, {allowDuplicate: true});
+
+    await caClient.BC.BTC.webhook.createTransactionConfirmationsWebHook(testUrl, testAddress, 1);
+    await caClient.BC.BTC.webhook.createTransactionConfirmationsWebHook(testUrl, testAddress, 1, {allowDuplicate: true});
+
+    await caClient.BC.BTC.webhook.createOmniConfirmedTransactionWebHook(testUrl, omniTestTx, 1);
+    await caClient.BC.BTC.webhook.createOmniConfirmedTransactionWebHook(testUrl, omniTestTx, 1, {allowDuplicate: true});
+
+    await caClient.BC.BTC.webhook.createOmniTransactionConfirmationsWebHook(testUrl, omniTestAddress, 1);
+    await caClient.BC.BTC.webhook.createOmniTransactionConfirmationsWebHook(testUrl, omniTestAddress, 1, {allowDuplicate: true});
+
+    await caClient.BC.BTC.webhook.createOmniTokenWebHook(testUrl, omniTestAddress, 1);
+    await caClient.BC.BTC.webhook.createOmniTokenWebHook(testUrl, omniTestAddress, 1, {allowDuplicate: true, everyConfirmation: true});
 
     await caClient.BC.BTC.webhook.listAllHooks();
     await caClient.BC.BTC.webhook.deleteAllMyWebhooks();