+++ tests/rest-apis/blockchains/XRP/index.js
@@ -1,0 +1,9 @@
+const XRP = require('./xrp.test');
+
+async function middleware(caClient, next, network = caClient.BC.XRP.NETWORKS.TESTNET) {
+    caClient.BC.XRP.switchNetwork(network);
+
+    await next(caClient);
+}
+
+module.exports = (caClient, network) => middleware(caClient, XRP, network);